# Add `LinearMap`, a map for very small sizes.