--- a/src/proto/mod.rs
+++ b/src/proto/mod.rs
@@ -1,3 +1,4 @@
 pub mod dlist;
+pub mod linear_map;
 pub mod par_vec;
 mod shootout;