This is a very simple `Map` implementation for small sizes. It scales horribly, due to the use of linear search. Only requires `PartialEq+Eq` for the key type.