@cgaebel 2015-01-07
I don't like this because it seems too trivial. Most operations can be done with a `Vec`, `swap_remove`, and `IteratorExt::find`.

Does it have any packages using it?