@tbu- 2015-01-07
`insert` and `remove` can't be trivially implemented, they need working around the borrow checker.

I implemented this because I need this for another library.