@reem 2015-01-07
@cgaebel I think it's basically as trivial as `SmallIntMap` or whatever it's called now, so if we have that we should have this (I think this is arguably more useful).