@cgaebel 2015-01-07
Neither collect-rs nor the rust standard library contain SmallIntMap.