@tbu- 2015-01-07
It's called `VecMap` now.