@cgaebel 2015-01-07
Why must this be a whole data structure instead of extension methods on Vec<(A, B)>? There's no special invariants to maintain, and it means you get the rest of the Vec infrastructure for free.