@tbu- 2015-01-07
The invariant that is maintained that there's only a single pair for each key. Also this makes it possible to furtherly optimize it using @Gankro's suggestion.