@cgaebel 2015-01-07
That's a good point. Maybe add a Deref impl for Vec<(A, B)>, and remove a bunch of the shared methods?