@tbu- 2015-01-07
You write optimizations in quotation marks, but that's exactly what you gain for not exposing everything – you also don't want to change your `HashMap` code because the underlying representation changed.