@cgaebel 2015-01-07
The vast majority of code in this file is just replicating things already provided by Vec. Most of it can be removed with a Deref impl. When the representation changes, feel free to remove the impl. It's not like collect-rs has a stability promise.