@stepancheg 2015-01-07
@cgaebel it is not about stability, it is about code readability. Wrapper struct it is important to emphasize that type is a map, not just utilities for vector.