@cgaebel 2015-01-07
The deref impl maintains the same invariants and greatly reduces code size.