@tbu- 2015-01-07
@cgaebel The changes you propose result in a performance loss of >10% in all benchmarks. I have no idea on why this is the case.

New:

    $ cargo bench linear_map
    [...]
    test proto::linear_map::bench::bench_get_middle_big               ... bench:       843 ns/iter (+/- 55)
    test proto::linear_map::bench::bench_get_middle_medium            ... bench:        97 ns/iter (+/- 3)
    test proto::linear_map::bench::bench_get_middle_small             ... bench:        12 ns/iter (+/- 2)
    test proto::linear_map::bench::bench_get_none_big                 ... bench:      1661 ns/iter (+/- 47)
    test proto::linear_map::bench::bench_get_none_medium              ... bench:       178 ns/iter (+/- 5)
    test proto::linear_map::bench::bench_get_none_small               ... bench:        18 ns/iter (+/- 1)
    test proto::linear_map::bench::bench_insert_big                   ... bench:    420411 ns/iter (+/- 10462)
    test proto::linear_map::bench::bench_insert_medium                ... bench:      5030 ns/iter (+/- 212)
    test proto::linear_map::bench::bench_insert_small                 ... bench:       160 ns/iter (+/- 51)
    test proto::linear_map::bench::bench_remove_insert_big            ... bench:    592218 ns/iter (+/- 15812)
    test proto::linear_map::bench::bench_remove_insert_medium         ... bench:      7494 ns/iter (+/- 362)
    test proto::linear_map::bench::bench_remove_insert_small          ... bench:       197 ns/iter (+/- 22)
    test proto::linear_map::bench::bench_remove_rev_insert_big        ... bench:    759637 ns/iter (+/- 48047)
    test proto::linear_map::bench::bench_remove_rev_insert_medium     ... bench:      9335 ns/iter (+/- 227)
    test proto::linear_map::bench::bench_remove_rev_insert_small      ... bench:       225 ns/iter (+/- 13)
     [...]

Old:

    $ cargo bench linear_map
    [...]
    test proto::linear_map::bench::bench_get_middle_big               ... bench:       706 ns/iter (+/- 16)
    test proto::linear_map::bench::bench_get_middle_medium            ... bench:        87 ns/iter (+/- 3)
    test proto::linear_map::bench::bench_get_middle_small             ... bench:         9 ns/iter (+/- 1)
    test proto::linear_map::bench::bench_get_none_big                 ... bench:      1406 ns/iter (+/- 151)
    test proto::linear_map::bench::bench_get_none_medium              ... bench:       155 ns/iter (+/- 16)
    test proto::linear_map::bench::bench_get_none_small               ... bench:        16 ns/iter (+/- 1)
    test proto::linear_map::bench::bench_insert_big                   ... bench:    290077 ns/iter (+/- 42532)
    test proto::linear_map::bench::bench_insert_medium                ... bench:      4104 ns/iter (+/- 361)
    test proto::linear_map::bench::bench_insert_small                 ... bench:       126 ns/iter (+/- 12)
    test proto::linear_map::bench::bench_remove_insert_big            ... bench:    453754 ns/iter (+/- 25046)
    test proto::linear_map::bench::bench_remove_insert_medium         ... bench:      6271 ns/iter (+/- 316)
    test proto::linear_map::bench::bench_remove_insert_small          ... bench:       170 ns/iter (+/- 24)
    test proto::linear_map::bench::bench_remove_rev_insert_big        ... bench:    620820 ns/iter (+/- 18436)
    test proto::linear_map::bench::bench_remove_rev_insert_medium     ... bench:      8411 ns/iter (+/- 237)
    test proto::linear_map::bench::bench_remove_rev_insert_small      ... bench:       188 ns/iter (+/- 6)
    [...]