@tbu- 2015-01-07
@cgaebel @Gankro This is a really simple data structure with a relatively straightforward implementation. Can't we just merge this? I'd like to experiment with the suggestion of two `Vec`s.